support epoll.

+++ t_utils.c
@@ -1,10 +1,10 @@
 #include <sys/time.h>
 #include <stdio.h>
 
-int get_now_time(void)
+unsigned int get_now_time(void)
 {
 	struct timeval tv;
 	
 	gettimeofday(&tv, NULL);
-	return tv.tv_usec / 1000 + tv.tv_sec * 1000;
+	return (unsigned int)(tv.tv_usec / 1000 + tv.tv_sec * 1000);
 }
